Move HAL code to separate file
also remove the link loading as I can't find a situation where it is used and the code needs refactoring before it'll even work.
